Refactor DukeGui
